Working with the methods of the portal, queue and user classes

Closes #106

See merge request iKono/acd-appservice!142

+++ acd_appservice/queue.py
@@ -1,9 +1,10 @@
 from __future__ import annotations
 
 import logging
-from typing import cast
+from typing import List, cast
 
 from mautrix.appservice import IntentAPI
+from mautrix.errors.base import IntentError
 from mautrix.types import RoomID, UserID
 from mautrix.util.logging import TraceLogger
 
@@ -14,7 +15,6 @@ from .user import User
 
 
 class Queue(DBQueue, MatrixRoom):
-
     room_id: RoomID
     name: str = ""
     description: str | None = None
@@ -33,7 +33,8 @@ class Queue(DBQueue, MatrixRoom):
         intent: IntentAPI = None,
     ):
         DBQueue.__init__(self, id=id, name=name, room_id=room_id, description=description)
-        MatrixRoom.__init__(self, room_id=self.room_id, intent=intent)
+        MatrixRoom.__init__(self, room_id=self.room_id)
+        self.main_intent = intent
         self.log = self.log.getChild(room_id)
 
     async def _add_to_cache(self) -> None:
@@ -53,10 +54,12 @@ class Queue(DBQueue, MatrixRoom):
         """It removes all members from the queue, leaves the queue, and deletes the queue"""
         members = await self.main_intent.get_joined_members(self.room_id)
 
+        if not members:
+            self.log.error(f"Unable to obtain members in the queue {self.room_id}")
+
         reason = "The queue will be removed"
 
         for member in members.keys():
-
             if self.main_intent.mxid == member:
                 continue
 
@@ -68,8 +71,9 @@ class Queue(DBQueue, MatrixRoom):
 
         memberships = await QueueMembership.get_by_queue(fk_queue=self.id)
 
-        for membership in memberships:
-            await membership.delete()
+        if memberships:
+            for membership in memberships:
+                await membership.delete()
 
         await self.delete()
 
@@ -89,6 +93,27 @@ class Queue(DBQueue, MatrixRoom):
             await QueueMembership.get_by_queue_and_user(fk_queue=self.id, fk_user=user.id)
 
         await self.save()
+
+    @classmethod
+    async def is_queue(cls, room_id: RoomID) -> bool:
+        """if a queue exists for the given room, return True, otherwise return False
+
+        Parameters
+        ----------
+        room_id : RoomID
+            The room ID of the queue.
+
+        Returns
+        -------
+            A boolean value.
+
+        """
+        queue = await cls.get_by_room_id(room_id=room_id, create=False)
+
+        if not queue:
+            return False
+
+        return True
 
     async def add_member(self, new_member: UserID):
         """If the config value for `acd.queue.user_add_method` is `join`, then join the user,
@@ -133,9 +158,107 @@ class Queue(DBQueue, MatrixRoom):
         await self.main_intent.set_room_name(room_id=self.room_id, name=new_name)
         await self.save()
 
+    async def get_agent_count(self) -> int:
+        """It returns the number of agents in the system
+
+        Returns
+        -------
+            The number of agents in the system.
+
+        """
+        agents = await self.get_agents() or []
+        return len(agents)
+
+    async def get_agents(self) -> List[User]:
+        """Get all the users in the channel, remove the bots, and return the remaining users
+
+        Returns
+        -------
+            A list of users
+
+        """
+        members = []
+
+        try:
+            members = await self.get_joined_users()
+        except IntentError as e:
+            self.log.error(e)
+
+        if not members:
+            return members
+
+        # remove bots from member list
+        return self.remove_not_agents(members)
+
+    def remove_not_agents(self, members: List[User]) -> List[User]:
+        """Removes non-agents from a list of users
+
+        Parameters
+        ----------
+        members : List[User]
+            List[User]
+
+        Returns
+        -------
+            A list of users that are agents.
+
+        """
+        only_agents: List[User] = []
+        if members:
+            # Removes non-agents
+            only_agents = [user for user in members if user.is_agent]
+
+        return only_agents
+
+    async def get_first_online_agent(self) -> User | None:
+        """It returns the first agent that is online in the room
+
+        Returns
+        -------
+            UserID
+
+        """
+        agents: List[User] = await self.get_agents()
+
+        if not agents:
+            self.log.debug(f"There's no agent in room: {self.room_id}")
+            return
+
+        for agent in agents:
+            # Switch between presence and agent operation login using config parameter
+            # to verify if agent is logged in
+
+            is_agent_online = await agent.is_online(queue_id=self.id)
+
+            if is_agent_online:
+                return agent
+
+    async def get_membership(self, agent: User) -> QueueMembership:
+        """It returns a QueueMembership object if the user is a member of the queue,
+        otherwise it returns None
+
+        Parameters
+        ----------
+        agent : User
+            User = The user that is being added to the queue
+
+        Returns
+        -------
+            A QueueMembership object
+
+        """
+
+        membership: QueueMembership = await QueueMembership.get_by_queue_and_user(
+            fk_user=agent.id, fk_queue=self.id, create=False
+        )
+
+        if not membership:
+            return
+
+        return membership
+
     @classmethod
     async def get_by_room_id(cls, room_id: RoomID, *, create: bool = True) -> Queue | None:
-
         try:
             return cls.by_room_id[room_id]
         except KeyError:
